requirements.txt
mysql-connector-python
flask
flasgger